pyproject.toml
[project]
name = "blog-manage"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "shizukani-cp", email = "shizukani.cp@gmail.com" }
]
dependencies = [
    "pyyaml>=6.0.2",
    "markdown>=3.7",
    "jinja2>=3.1.4",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/blog_manage"]

[dependency-groups]
dev = [
    "build>=1.2.2.post1",
    "twine>=6.1.0",
]
